requirements.txt
flask
flask-cors
openai
langchain
langchain-community
langchain-openai
faiss-cpu
python-dotenv
